Added schema V5

+++ Kafka/RideShareProducer/src/main/java/com/insight/RideShareProducerBase.java
@@ -59,7 +59,7 @@ public abstract class RideShareProducerBase <Ride> {
                 final String[] taxiTrip = line.split(cvsSplitBy, -18);
 
 //                String uniqueID = UUID.randomUUID().toString();
-                String uniqueID = Instant.now().getEpochSecond()+String.valueOf(UUID.randomUUID().getMostSignificantBits() & Long.MAX_VALUE );
+                String uniqueID = Instant.now().getEpochSecond()+String.valueOf(UUID.randomUUID().getMostSignificantBits() & Integer.MAX_VALUE );
 
                 if (i > 0 && !line.contains("NULL")) {
                     buildRecord( taxiTrip );
